requirements-linux.txt
scapy-python3
readline
psutil
ipaddress
colorama